// src/rot/display/canvas.ts
/**!
 * Copyright (c) 2012-now(), Ondrej Zara
 * All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without modification,
 * are permitted provided that the following conditions are met:
 *
 * 	* Redistributions of source code must retain the above copyright notice,
 * 	  this list of conditions and the following disclaimer.
 * 	* Redistributions in binary form must reproduce the above copyright notice,
 * 	  this list of conditions and the following disclaimer in the documentation
 * 	  and/or other materials provided with the distribution.
 * 	* Neither the name of Ondrej Zara nor the names of its contributors may be used
 * 	  to endorse or promote products derived from this software without specific
 * 	  prior written permission.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
 * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
 * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
 * IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
 * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
 * OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
 * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */

import Backend from "./backend";
import { DisplayOptions } from "./types";

export default abstract class Canvas extends Backend {
    _ctx: CanvasRenderingContext2D;

    constructor() {
        super();
        this._ctx = document.createElement("canvas").getContext("2d") as CanvasRenderingContext2D;
    }

    getContainer() { return this._ctx.canvas; }

    setOptions(opts: DisplayOptions) {
        super.setOptions(opts);

        const style = (opts.fontStyle ? `${opts.fontStyle} ` : "");
        const font = `${style} ${opts.fontSize}px ${opts.fontFamily}`;
        this._ctx.font = font;
        this._updateSize();

        this._ctx.font = font;
        this._ctx.textAlign = "center";
        this._ctx.textBaseline = "middle";
    }

    clear() {
        this._ctx.fillStyle = this._options.bg;
        this._ctx.fillRect(0, 0, this._ctx.canvas.width, this._ctx.canvas.height);
    }

    eventToPosition(x: number, y: number): [number, number] {
        const canvas = this._ctx.canvas;
        const rect = canvas.getBoundingClientRect();
        x -= rect.left;
        y -= rect.top;

        x *= canvas.width / rect.width;
        y *= canvas.height / rect.height;

        if (x < 0 || y < 0 || x >= canvas.width || y >= canvas.height) { return [-1, -1]; }

        return this._normalizedEventToPosition(x, y);
    }

    abstract _normalizedEventToPosition(x: number, y: number): [number, number];
    abstract _updateSize(): void;
}
